medical_image_classifier_full/requirements.txt
streamlit
tensorflow
numpy
matplotlib
Pillow
